Remove all uses of allocRaw
We used to combine multiple allocations so that we could return a single
pointer that the caller could free, instead of having to track multiple
allocations. However, we now have a helper for doing exactly that.

This will cause a slight decrease in performance, but the right solution
to that is implementing a better allocator, not doing error-prone
pointer arithmetic.